Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 int main(){
-    int num;
+    float num;
     printf("You entered: %.2f");
     scanf("%f", &num);
     return 0;
